pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["Excidion <36995046+Excidion@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
polars = "^0.20.25"
plotly = "^5.22.0"
pandas = "^2.2.2"
xlsx2csv = "^0.8.2"
pyyaml = "^6.0.1"
nbformat = "^5.10.4"
streamlit = "^1.34.0"
scikit-learn = "^1.4.2"
pyarrow = "^16.0.0"
dash-bio = "^1.0.2"
umap-learn = "^0.5.6"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.4"
ruff = "^0.4.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
